backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
redis==5.0.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
PyMuPDF==1.23.8
python-docx==1.1.0
beautifulsoup4==4.12.2
ibm-watsonx-ai==0.2.0
python-dotenv==1.0.0
